feat: fetchWithoutAuth 함수 추가 및 응답 처리 로직을 handleResponse 함수로 분리
